BOOK-75 auth.ts refatorado
